test(accordion): migrate multiple test
