+++ core/src/components/accordion/test/multiple/accordion.e2e.ts
@@ -1,0 +1,32 @@
+import { expect } from '@playwright/test';
+import { configs, test } from '@utils/test/playwright';
+
+test.describe('accordion: multiple', () => {
+  configs({ direction: ['ltr'] }).forEach(({ config, screenshot, title }) => {
+    test(title('should update value and visually expand items'), async ({ page }) => {
+      await page.goto(`/src/components/accordion/test/multiple`, config);
+      const accordionGroup = page.locator('ion-accordion-group');
+      const diningHeader = page.locator('ion-accordion[value="dining"] ion-item[slot="header"]');
+      const attractionsHeader = page.locator('ion-accordion[value="attractions"] ion-item[slot="header"]');
+
+      await expect(accordionGroup).toHaveJSProperty('value', 'attractions');
+
+      await expect(accordionGroup).toHaveScreenshot(screenshot('accordion-one-open'));
+
+      await diningHeader.click();
+      await page.waitForChanges();
+
+      await expect(accordionGroup).toHaveJSProperty('value', ['attractions', 'dining']);
+
+      await expect(accordionGroup).toHaveScreenshot(screenshot('accordion-two-open'));
+
+      await diningHeader.click();
+      await attractionsHeader.click();
+      await page.waitForChanges();
+
+      await expect(accordionGroup).toHaveJSProperty('value', []);
+
+      await expect(accordionGroup).toHaveScreenshot(screenshot('accordion-zero-open'));
+    });
+  });
+});
